requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
